Extracted consumption reserve tests
